requirements.txt
fastapi
uvicorn[standard]
jinja2
pydantic
python-multipart
PyYAML
rich
portia-sdk-python[google]
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib


